backend/requirements.txt
torch
torchvision
flwr==1.8.0
web3==6.15.1
numpy
requests
pycryptodome
